+++ src/main/java/shahzam/exception/DataIntegrityException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class DataIntegrityException extends ShahzamExceptions {
+    public DataIntegrityException() {
+        super("shahzam.utils.Storage file integrity compromised :(");
+    }
+}
