+++ src/main/java/shahzam/exception/EmptyTaskDescriptionException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class EmptyTaskDescriptionException extends ShahzamExceptions {
+    public EmptyTaskDescriptionException(String message) {
+        super(message);
+    }
+}
