+++ src/main/java/shahzam/exception/IllegalFormatException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class IllegalFormatException extends ShahzamExceptions {
+    public IllegalFormatException(String format) {
+        super("Please follow this format:\n  " + format);
+    }
+}
