+++ src/main/java/shahzam/exception/InvalidDeadlineFormatException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class InvalidDeadlineFormatException extends ShahzamExceptions {
+    public InvalidDeadlineFormatException(String message) {
+        super(message);
+    }
+}
