+++ src/main/java/shahzam/exception/InvalidEventFormatException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class InvalidEventFormatException extends ShahzamExceptions {
+    public InvalidEventFormatException(String message) {
+        super(message);
+    }
+}
