+++ src/main/java/shahzam/exception/InvalidTaskNumberException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class InvalidTaskNumberException extends ShahzamExceptions {
+    public InvalidTaskNumberException(String message) {
+        super(message);
+    }
+}
