+++ src/main/java/shahzam/exception/ShahzamExceptions.java
@@ -1,0 +1,8 @@
+package shahzam.exception;
+
+public class ShahzamExceptions extends Exception {
+    public ShahzamExceptions(String message) {
+        super(message);
+    }
+}
+
