+++ src/main/java/shahzam/exception/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package shahzam.exception;
+
+public class UnknownCommandException extends ShahzamExceptions {
+    public UnknownCommandException() {
+        super("Sorry but my knowledge does not have such command.");
+    }
+}
