+++ src/main/java/shahzam/task/Deadline.java
@@ -1,3 +1,7 @@
+package shahzam.task;
+
+import shahzam.utils.DateTimeFormatUtils;
+
 import java.time.LocalDateTime;
 
 public class Deadline extends Task {
