+++ src/main/java/shahzam/task/Event.java
@@ -1,3 +1,7 @@
+package shahzam.task;
+
+import shahzam.utils.DateTimeFormatUtils;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
