+++ src/main/java/shahzam/task/Task.java
@@ -1,3 +1,5 @@
+package shahzam.task;
+
 public class Task {
     private String description;
     private boolean isDone;
